requirements.txt
python-telegram-bot==13.10
requests==2.27.1
urlextract==1.5.0
urllib3==1.26.8
youtube-dl==2021.12.17
psycopg2==2.9.3
